ip-reporter/requirements.txt
pyyaml
rumps  # macOS menu bar apps
pytest  # for testing
